Sql Injection-Insecure app/requirements.txt
# insecure/requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.27
